fixup??? fscache: teach fscache to use NtQueryDirectoryFile
Sadly, we include too many variants of too many header files that all
define `UNICODE_STRING`: `winternl.h`, `NTSecAPI.h`, `ntdef.h`. We tried
to define `UNICODE_STRING` so that we can declare `NtQueryDirectory()`
with a minimal set of type definitions. But this causes compile errors,
now that we need to have `RtlGenRandom` declared.

Let's just include `ntdef.h` and be done with it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>